+++ packages/vue2/src/web-widget.ts
@@ -114,21 +114,26 @@ export /*#__PURE__*/ function defineWebWidget(
   options: DefineWebWidgetOptions
 ) {
   console.warn(`defineWebWidget of vue2 is an experimental feature.`);
-
-  options.renderTarget = "light"; // TODO shadow
-  options.loading = "lazy";
-  return /*#__PURE__*/ defineComponent({
+  return defineComponent({
     name: "WebWidgetSuspense",
     props: {
+      fallback: {
+        type: Object as PropType<Component>,
+      },
+      loading: {
+        type: String as PropType<WebWidgetContainerOptions["loading"]>,
+        default: options.loading ?? "lazy",
+      },
       renderStage: {
         type: String as PropType<WebWidgetContainerOptions["renderStage"]>,
         default: options.renderStage,
       },
-      fallback: {
-        type: Object as PropType<Component>,
+      renderTarget: {
+        type: String as PropType<WebWidgetContainerOptions["renderTarget"]>,
+        default: options.renderTarget ?? "light",
       },
     },
-    setup({ fallback, renderStage }, { slots }) {
+    setup({ fallback, loading, renderStage, renderTarget }, { slots }) {
       // eslint-disable-next-line react-hooks/rules-of-hooks
       const data = useAttrs() as WebWidgetContainerOptions["data"];
 
@@ -138,7 +143,9 @@ export /*#__PURE__*/ function defineWebWidget(
             ...options,
             data,
             loader,
+            loading,
             renderStage,
+            renderTarget,
 
             // -----
             fallback,